+++ PokemonApp/Models/PokemonBaseDetailModel.swift
@@ -1,0 +1,17 @@
+//
+//  PokemonBaseDetailModel.swift
+//  PokemonApp
+//
+//  Created by Byron Chavarría on 7/8/20.
+//  Copyright © 2020 Byron Chavarría. All rights reserved.
+//
+
+import UIKit
+
+struct PokemonBaseDetailModel {
+    let pokemonImageUrl: String
+    let pokemonName: String
+    let baseHappiness: String
+    let captureRate: String
+    let generation: String
+}