+++ PokemonApp/Models/Response/PokemonResponse.swift
@@ -8,18 +8,12 @@
 
 import Foundation
 
-final class PokemonResponse: Decodable {
+struct PokemonResponse: Codable {
     let name: String?
     let url: String?
     
     enum CodingKeys: String, CodingKey {
         case name
         case url
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        name = try container.decodeIfPresent(String.self, forKey: .name)
-        url = try container.decodeIfPresent(String.self, forKey: .url)
     }
 }