+++ PokemonApp/ViewController/PokemonDetailViewController/PokemonDetailViewControllerDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  PokemonDetailViewControllerDelegate.swift
+//  PokemonApp
+//
+//  Created by Byron Chavarría on 7/8/20.
+//  Copyright © 2020 Byron Chavarría. All rights reserved.
+//
+
+import Foundation
+
+protocol PokemonDetailViewControllerDelegate: class {
+    func PokemonDetailViewControllerDidSelectPokemon()
+}